+++ buffer/buffer_test.go
@@ -29,7 +29,7 @@ import (
 )
 
 func TestBufferWrites(t *testing.T) {
-	buf := New()
+	buf := NewPool().Get()
 
 	tests := []struct {
 		desc string
@@ -69,7 +69,7 @@ func BenchmarkBuffers(b *testing.B) {
 	str := strings.Repeat("a", 1024)
 	slice := make([]byte, 1024)
 	buf := bytes.NewBuffer(slice)
-	custom := New()
+	custom := NewPool().Get()
 	b.Run("ByteSlice", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			slice = append(slice, str...)
