+++ zapcore/json_encoder_bench_test.go
@@ -25,7 +25,6 @@ import (
 	"testing"
 	"time"
 
-	"go.uber.org/zap/internal/bufferpool"
 	. "go.uber.org/zap/zapcore"
 )
 
@@ -56,7 +55,7 @@ func BenchmarkZapJSON(b *testing.B) {
 				Message: "fake",
 				Level:   DebugLevel,
 			}, nil)
-			bufferpool.Put(buf)
+			buf.Free()
 		}
 	})
 }
